+++ epyk/__init__.py
@@ -4,5 +4,6 @@
 
 from epyk.core import Page as Rpt
 from epyk.core.data import events
+from epyk.core.css.styles.attributes import CssInline
 
 Page = Rpt.Report